Move test data creation to shared module
