+++ test/test_helper.rb
@@ -24,6 +24,7 @@ require 'integration/api/read_api_test_suite'
 require 'integration/api/write_api_test_suite'
 require 'rdf_test_cases'
 require 'rack_test_cookie_jar_extensions'
+require 'single_page_test_utils'
 
 Minitest::Reporters.use! [Minitest::Reporters::DefaultReporter.new(fast_fail: true,
                                                                    color: true,
